matrix_bot/requirements.txt
matrix-nio[e2e]>=0.24
httpx>=0.26
python-dotenv>=1.0
